Improve StringData::Make(SharedVariant*)
The StringData inside a SharedVariant caches the hash for the
shared string, but we wait until the first time it is accessed to load
it into the request-local string.  Since we're already loading things
out of the shared memory at creation time, just eagerly load it there.
(It might be zero if no one has hashed it yet: we update it racily
from whichever request hashes it first.)  This diff also shuffles some
code around to avoid extra checks for MaxSize, and removes a few
redundant loads that gcc couldn't remove in some Make functions.  Also
move the code to shared-variant.o, and inline the fast path into its
only caller.

Reviewed By: @edwinsmith

Differential Revision: D960931