Refactor of IssetEmptyElem
Cleaning up the last Elem function of member-operations:

As far as I (or gcc) could tell, the isObj template parameter wasn't
used so I nuked it.

ElemIssetEmptyString was begging for a fast path when the key is an
Int64, so I obliged.  I think there's a bit more fat there but I'll
get to it in a later diff.

Reviewed By: @edwinsmith

Differential Revision: D962556