Sending the breakpoint list from the client to the proxy clobbers the data structures that track which breakpoints are temporarily disabled
Whenever the client sends a command to the proxy that includes a list of breakpoints, the proxy clobbers its list of breakpoints with the new list received from the client. Unfortunately it also uses its list of breakpoints to track which breakpoints are temporarily disabled because the debugger is currently stopped there. As a result, just listing the active breakpoints while stopped at a break point will cause the next continue command to hit the current break point if more byte codes have yet to be processed.

Reviewed By: @mikemag

Differential Revision: D962443