Cleanup some low_malloc issues
I noticed that the first page of low memory never gets hugified,
because the arena gets setup before main, and we don't know what to
mark huge until we've initialized the runtime options.

By setting the high water mark to 1 below the start of the low memory
arena, and only updating it when we *are* marking pages huge, we can
ensure that the right pages get marked.

I also replaced some jemalloc magic that I had never previously understood
with a simpler (to me) equivalent.

Reviewed By: @jdelong

Differential Revision: D967121