Clean up IsType hacks for objects and resources
This was unlocked by separating resources and objects. is_object now
behaves normally (it's just a check for m_type == KindOfObject, instead
of needing to exclude resources), so all the special logic in codegen
and simplifier can go away.

I also replaced the string-type logic with an assert, and added an
equivalent one for array types; the special logic wasn't necessary in
practice because we always pass Type::Str as the tested type. This is
just future-proofing.

Type::Res doesn't need any special treatment because there's no opcode
for is_resource().

Reviewed By: @ottoni

Differential Revision: D970744