Remove view() method from collections, also remove Vector::sort()
The view() method has been phased out in favor of lazy() (which does the
same thing but is more obvious to developers). Let's remove it.

Also, the Vector::sort() method has been deprecated for a while and was
abandoned in favor of making the regular sort builtin functions work with
collections, let's get rid of that too while we're at it.

Reviewed By: drussi

Differential Revision: D979375