Detect and report stack imbalance in TraceProfiler
TraceProfiler maintains a log of every function call and return. By adding the function name to return events we can ensure that a return matches a call when post-processing the log. Too keep the log size small I've stolen a bit out of the memory counter for the trace samples since it's not realistic that we'd need a full 64 bits there.

I've added a little bit of logic to detect a missing return and correct the stack imbalance to keep the rest of the report accurate.

Reviewed By: @swtaarrs

Differential Revision: D983012