Remove the trace setter in gen()
Instructions ultimately want a block to jump to, not a trace,
and the trace setter is in the way of untangling some header
dependencies I have in a separate diff.  This was a big enough
change to factor out separately.

It's essentially a mechanical change, although I changed the
helper functions that create traces to return the first block
instead, for convenience.

Reviewed By: @swtaarrs

Differential Revision: D984304