check subvert function depend on context Class
cf scannel indication:
It should be transparent but in your implementation, it isn't, even with version #2.

Example: Let's say A::foo and B::private_bar are both registered autoload handlers. If I'm autoloading using A::foo, A::foo would still be able to directly call B::private_bar [not through autoload, as a direct method call] even though it is private because it is registered as an autoload handler.