Remove bytecode invariant about fault funclet uniqueness
We used to need this because when executing an Unwind
instruction we tried to find a PC to re-raise the fault at based on
the offset of the fault funclet.  This didn't quite work (adjacent
fault regions weren't handled correctly), and was changed in D697005,
so I think we can drop this requirement now.

Reviewed By: @edwinsmith

Differential Revision: D1000915