avoid undefined __builtin_clz(0) behavior in nextPower2
HPHP::Util::nextPower2(1) ends up calling __builtin_clz(0),
which has undefined behavior.  In practice it can return any result.
None of the current callers seem to trigger the bug, but in addition to
fixing the bug this diff also reduces the compiled size and increases
the performance of the method.

Also, the attempt to use an arithmetic exception instead of an assert
doesn't seem to be working, so I replaced it with a regular assert.

Reviewed By: @andralex

Differential Revision: D1007106